# backend/OmnikeeperOnlineInboundAdapter/Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Landscape Registry REST API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:44378"
    }
  ],
  "paths": {
...
    "/api/v{version}/Attribute/getMergedAttributesWithName": {
      "get": {
        "tags": [
          "Attribute"
        ],
        "summary": "",
        "operationId": "GetMergedAttributesWithName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "description": ""
            }
          },
          {
            "name": "layerIDs",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              },
              "description": ""
            }
          },
          {
            "name": "atTime",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string",
              "description": "",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CIAttributeDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CIAttributeDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CIAttributeDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/v{version}/Attribute/getMergedAttributes": {
      "get": {
        "tags": [
          "Attribute"
        ],
        "summary": "",
        "operationId": "GetMergedAttributes",
        "parameters": [
          {
            "name": "ciids",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              },
              "description": ""
            }
          },
          {
            "name": "layerIDs",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              },
              "description": ""
            }
          },
          {
            "name": "atTime",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string",
              "description": "",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CIAttributeDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CIAttributeDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CIAttributeDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/v{version}/Attribute/bulkReplaceAttributesInLayer": {
      "post": {
        "tags": [
          "Attribute"
        ],
        "summary": "bulk replace all attributes in specified layer",
        "operationId": "BulkReplaceAttributesInLayer",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCIAttributeLayerScopeDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCIAttributeLayerScopeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCIAttributeLayerScopeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCIAttributeLayerScopeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttributeValueType": {
        "enum": [
          "Text",
          "MultilineText",
          "Integer",
          "JSON",
          "YAML"
        ],
        "type": "string"
      },
      "AttributeValueDTO": {
        "required": [
          "isArray",
          "type",
          "values"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/AttributeValueType"
          },
          "isArray": {
            "type": "boolean"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AttributeState": {
        "enum": [
          "New",
          "Changed",
          "Removed",
          "Renewed"
        ],
        "type": "string"
      },
      "CIAttributeDTO": {
        "required": [
          "ciid",
          "id",
          "name",
          "state",
          "value"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "readOnly": true
          },
          "value": {
            "$ref": "#/components/schemas/AttributeValueDTO"
          },
          "ciid": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "state": {
            "$ref": "#/components/schemas/AttributeState"
          }
        }
      },
      "FragmentDTO": {
        "required": [
          "ciid",
          "name",
          "value"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/AttributeValueDTO"
          },
          "ciid": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "BulkCIAttributeLayerScopeDTO": {
        "required": [
          "fragments",
          "layerID",
          "namePrefix"
        ],
        "type": "object",
        "properties": {
          "namePrefix": {
            "type": "string"
          },
          "layerID": {
            "type": "integer",
            "format": "int64"
          },
          "fragments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FragmentDTO"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "authorizationUrl": "https://host.docker.internal:8443/auth/realms/landscape/protocol/openid-connect/auth",
            "tokenUrl": "https://host.docker.internal:8443/auth/realms/landscape/protocol/openid-connect/token",
            "scopes": { }
          }
        }
      }
    }
  }
}